+++ include/draw.hpp
@@ -1,5 +1,3 @@
 #pragma once
 
 void drawJuliaFatouImage();
-
-void drawImgui();
