+++ include/gui.hpp
@@ -1,3 +1,8 @@
 #pragma once
 
-void initWindows();
+#include <GLFW/glfw3.h>
+
+void initSettingsWindow();
+void setUpImgui(GLFWwindow* window, const char* glsl_version);
+void drawImgui();
+void cleanUpImgui();
